+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/levelgen/feature/configured/BWGVanillaConfiguredFeatures.java
@@ -2,9 +2,12 @@ package net.potionstudios.biomeswevegone.world.level.levelgen.feature.configured
 
 import net.minecraft.core.HolderSet;
 import net.minecraft.data.worldgen.features.FeatureUtils;
+import net.minecraft.data.worldgen.features.VegetationFeatures;
 import net.minecraft.data.worldgen.placement.PlacementUtils;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.util.random.SimpleWeightedRandomList;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.RandomPatchConfiguration;
@@ -12,6 +15,7 @@ import net.minecraft.world.level.levelgen.feature.configurations.SimpleBlockConf
 import net.minecraft.world.level.levelgen.feature.configurations.SimpleRandomFeatureConfiguration;
 import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProvider;
 import net.minecraft.world.level.levelgen.feature.stateproviders.NoiseThresholdProvider;
+import net.minecraft.world.level.levelgen.feature.stateproviders.WeightedStateProvider;
 import net.minecraft.world.level.levelgen.synth.NormalNoise;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import net.potionstudios.biomeswevegone.world.level.block.BWGBlocks;
@@ -19,6 +23,19 @@ import net.potionstudios.biomeswevegone.world.level.block.BWGBlocks;
 import java.util.List;
 
 public class BWGVanillaConfiguredFeatures {
+
+    public static final ResourceKey<ConfiguredFeature<?, ?>> FLOWER_DEFAULT = ConfiguredFeaturesUtil.createConfiguredFeature(
+            "vanilla/flower_default",
+            Feature.FLOWER,
+            () -> VegetationFeatures.grassPatch(new WeightedStateProvider(SimpleWeightedRandomList.<BlockState>builder()
+                    .add(BWGBlocks.FLOWER_PATCH.get().defaultBlockState(), 2)
+                    .add(BWGBlocks.CLOVER_PATCH.get().defaultBlockState(), 2)
+                    .add(BWGBlocks.WHITE_ANEMONE.getBlockState(), 1)
+                    .add(BWGBlocks.PINK_ANEMONE.getBlockState(), 1)
+                    .add(BWGBlocks.ORANGE_DAISY.getBlockState(), 1)
+                    .add(BWGBlocks.HORSEWEED.getBlockState(), 1)
+                    .add(BWGBlocks.ROSE.getBlockState(), 1)
+                    .build()), 64));
 
     public static final ResourceKey<ConfiguredFeature<?, ?>> FLOWER_PLAINS = ConfiguredFeaturesUtil.createConfiguredFeature(
             "vanilla/flower_plains",